Rename procfile -> processfile
Avoid conflicts with `Procfile`
