+++ processfile/doc.go
@@ -1,7 +1,7 @@
-// Package procfile allows for reading, modifying, and writing Procfile format
+// Package processfile allows for reading, modifying, and writing Procfile format
 // files.
 //
 // It does not preserve the ordering of definitions within a Procfile.
 //
 // See: https://devcenter.heroku.com/articles/procfile
-package procfile
+package processfile
